Update the warmshowers_site module to deal gracefully with htmlmail
It turns out that some implementation of hook_mail() return ['body']
as a string instead of as an array. We get blasted in that situation.
This commit makes warmshowers_site respect that situation (and not add
the footer) and also adds a hook_update_N() that sets htmlmail to a higher
weight.