Revert "Fix broken hover style on "mark as spam" button" for #132
This reverts commit a8448ebcf8c9e6eb12e0c31e664835ac35a57cc0.

I attempted to do the "right thing" where both @joshmiller and @illmasterc
had also attempted to do the right thing. However... I got it working again.
The "Mark as Spam" buttons and "Mark Obsolete" are probably not what we want
but not nearly as important as having the regular linkbuttons.

Conflicts:
	sites/all/themes/dev/warmshowers_zen/css/elements/linkbutton.css