Add views templates for dual feedback view
Discussed with randy a transition to one page
with both feedback received and given for a user.
These templates provide clarity using the member's
name and link down the page with an anchor to also
help with usability.  I've exported the view
and loaded it into the live site for testing as it
will not be live immediately, but will require
some link changes