Changed style of recommendation field labels and items
Current labels and items on `/recommendation/feedback-xxx` or stacked
on top of each other. When a profile picture is present, the response
to `overall experience` appears below the picture whereas the label
appears to the right of the picture.

This CSS tweek simply floats the label so that the item is inlined and
looks more natural and readable. It should work on all browsers.