Add testing capability to reminder function, and form/handler to run tests.
Modify the function that queues reminders/pester emails for unavailable
users to allow for easier testing. Add form elements on the /admin/user/wsuser
page, and a submit handler, to run the tests. Info on users that would
receive reminder emails is displayed in the message area, and actual
emails are sent to the specified test email address. RE: Pull Request  #364