Fixes #924 search failure and improves #174 (best-effort search when autocomplete ignored)
It turns out that there had been an old attempt at #174
that was still in the code and also didn't get upgrade properly.
Since it was only hit in the case where somebody didn't wait for the
autocomplete, we didn't see it often.

I have only left city, province search in there for the US, as
most countries don't have a usable province code in our db.
Even canada has numeric province codes in geonames_adm1, which
is baffling.